+++ app/src/main/java/com/olup/notable/classes/DrawCanvas.kt
@@ -23,6 +23,7 @@ import com.olup.notable.db.selectImage
 import com.olup.notable.db.selectImagesAndStrokes
 import com.olup.notable.utils.History
 import com.olup.notable.utils.Operation
+import com.olup.notable.utils.uriToBitmap
 import com.onyx.android.sdk.api.device.epd.EpdController
 import com.onyx.android.sdk.data.note.TouchPoint
 import com.onyx.android.sdk.pen.RawInputCallback
