+++ app/src/main/java/com/olup/notable/components/ToolbarMenu.kt
@@ -22,6 +22,12 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Popup
 import androidx.compose.ui.window.PopupProperties
 import androidx.navigation.NavController
+import com.olup.notable.utils.copyPagePngLinkForObsidian
+import com.olup.notable.utils.exportBook
+import com.olup.notable.utils.exportBookToPng
+import com.olup.notable.utils.exportPage
+import com.olup.notable.utils.exportPageToJpeg
+import com.olup.notable.utils.exportPageToPng
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
