+++ app/src/main/java/com/olup/notable/utils/draw.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.toOffset
 import com.olup.notable.db.Image
 import com.olup.notable.db.Stroke
+import com.olup.notable.utils.uriToBitmap
 import com.onyx.android.sdk.data.note.ShapeCreateArgs
 import com.onyx.android.sdk.data.note.TouchPoint
 import com.onyx.android.sdk.pen.NeoBrushPen
