+++ app/src/main/java/com/olup/notable/utils/page.kt
@@ -1,4 +1,4 @@
-package com.olup.notable
+package com.olup.notable.utils
 
 
 import android.content.ClipData
@@ -17,9 +17,15 @@ import android.provider.MediaStore
 import androidx.compose.ui.unit.IntOffset
 import androidx.core.content.ContextCompat
 import androidx.core.content.FileProvider
+import com.olup.notable.SCREEN_HEIGHT
+import com.olup.notable.SCREEN_WIDTH
+import com.olup.notable.TAG
 import com.olup.notable.db.BookRepository
 import com.olup.notable.db.PageRepository
 import com.olup.notable.db.Stroke
+import com.olup.notable.drawBg
+import com.olup.notable.drawImage
+import com.olup.notable.drawStroke
 import io.shipbook.shipbooksdk.Log
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
